refactor: :truck: Changing routes.
Routes are changed for some variables that are no longer located in the file's location.
